Port over BlinkenJS page, behaviour and styles
To make it easier to iterate the 2nd line dashboard, let's keep
the BlinkenJS frame code in this repository, so that we can make
changes to the dashboard and see them reflected immediately.

This code is taken from https://github.com/alphagov/govuk-secondline-blinken.

Note that the Chrome extension hasn't been copied over, but it
can still be retrieved from the old repo if needed, and is unlikely
to be useful especially now we've replatformed.