Set root to redirect to predefined `/frame-splits` URL
This way, devs only need to bookmark the root URL of the app, and
will automatically get any future changes made to the app. The
previous approach was to bookmark the frame splits URL itself,
but that made updating the dashboard a manual task for each
developer.

This new layout also has room for a 'summary' frame, which
provides links to useful resources and adds context as to why
we still need links to Icinga alerts (as some are still relevant
for the time being).